style: Hex 코드 tsconfig 지정색으로 변경
